imported_saved_backup/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
qrcode==7.4.2
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
openai>=1.12.0
psycopg2-binary==2.9.7
